Made TileMetaInfoMgr manage the global list of tilesets.
BuildingEd had its own list of tilesets in TMXConfig.txt.
TileMetaInfoMgr had its own list of tilesets in TileMetaInfo.txt.
PathGenerators.txt (in another branch) had its own list of tilesets.
Now only Tilesets.txt has the global list of tilesets.

TileMetaInfoDialog gained an entry & button for choosing the Tiles directoryl
BuildingPreferences and BuildingPreferencesDialog lost the tiles directory.

BuildingTilesDialog doesn't add/remove tilesets itself, it displays the
TileMetaInfoDialog.

+++ src/tiled/mainwindow.cpp
@@ -532,6 +532,7 @@ MainWindow::~MainWindow()
     ToolManager::deleteInstance();
 #ifdef ZOMBOID
     MapManager::deleteInstance();
+    TileMetaInfoMgr::deleteInstance();
 #endif
     TilesetManager::deleteInstance();
     DocumentManager::deleteInstance();
@@ -1290,23 +1291,25 @@ void MainWindow::showBuildingEditor()
 void MainWindow::tilesetMetaInfoDialog()
 {
     TileMetaInfoMgr *mgr = TileMetaInfoMgr::instance();
-    if (!mgr->readTxt()) {
-        QMessageBox::warning(this, tr("It's no good, Jim!"), mgr->errorString());
-        TileMetaInfoMgr::deleteInstance();
-        return;
+    if (!mgr->hasReadTxt()) {
+        if (!mgr->readTxt()) {
+            QMessageBox::warning(this, tr("It's no good, Jim!"),
+                                 tr("%1\n(while reading %2)")
+                                 .arg(mgr->errorString())
+                                 .arg(mgr->txtName()));
+            TileMetaInfoMgr::deleteInstance();
+            return;
+        }
+        PROGRESS progress(tr("Loading Tilesets.txt tilesets"));
+        mgr->loadTilesets();
     }
-
-    mgr->loadTilesets();
 
     TileMetaInfoDialog dialog(this);
     dialog.exec();
 
     if (!mgr->writeTxt()) {
         QMessageBox::warning(this, tr("It's no good, Jim!"), mgr->errorString());
-        TileMetaInfoMgr::deleteInstance();
     }
-
-    TileMetaInfoMgr::deleteInstance();
 }
 #endif
